LayoutTests/fast/dom/attribute-case-sensitivity2-expected.txt

case sensitive?	tag	namespace	id	attribute	namespace	first value	second value
Parser Tests:
true	p	http://www.w3.org/1999/xhtml	test1	case	undefined	first	second
true	foobar	http://www.w3.org/1999/xhtml	test2	case	undefined	first	second
true	svg	http://www.w3.org/2000/svg	test3	case	undefined	first	second
Javascript Tests:
true	p	null	test4	testattr	undefined	first	second
true	p	null	test4	namespacetest	http://www.example.com/	first	second
true	p	null	test4	href	http://www.w3.org/1999/xlink	first	second
true	foobar	null	test5	testattr	undefined	first	second
true	foobar	null	test5	namespacetest	http://www.example.com/	first	second
true	foobar	null	test5	href	http://www.w3.org/1999/xlink	first	second
true	barbar	http://www.example.com/	test6	testattr	undefined	first	second
true	barbar	http://www.example.com/	test6	namespacetest	http://www.example.com/	first	second
true	barbar	http://www.example.com/	test6	href	http://www.w3.org/1999/xlink	first	second
true	svg	http://www.w3.org/2000/svg	test7	testattr	undefined	first	second
true	svg	http://www.w3.org/2000/svg	test7	namespacetest	http://www.example.com/	first	second
true	svg	http://www.w3.org/2000/svg	test7	href	http://www.w3.org/1999/xlink	first	second
